% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMSstrings.R
\name{parse_mods}
\alias{parse_mods}
\title{parse_mods
Extract metadata from an annotated peptide sequence from Spectronaut}
\usage{
parse_mods(seqs, format = "Spectronaut")
}
\arguments{
\item{seqs}{A character vector of peptide sequences}

\item{format}{A character value specifying the format of our seqs (not currently implemented - function expects Spectronaut format for now)}
}
\value{
A 'protein_mods` object with processed sequences
}
\description{
parse_mods
Extract metadata from an annotated peptide sequence from Spectronaut
}
